refactor: move to component package
